Struct dialoguer::Input

pub struct Input<'a, T> { /* private fields */ }

Renders an input prompt.

Example usage

use dialoguer::Input;

let input : String = Input::new()
    .with_prompt("Tea or coffee?")
    .with_initial_text("Yes")
    .default("No".into())
    .interact_text()?;

It can also be used with turbofish notation:

let input = Input::<String>::new()
    .interact_text()?;

Implementations

impl<T> Input<'_, T>

pub fn new() -> Self

Creates an input prompt.

pub fn with_prompt<S: Into<String>>(&mut self, prompt: S) -> &mut Self

Sets the input prompt.

pub fn report(&mut self, val: bool) -> &mut Self

Indicates whether to report the input value after interaction.

The default is to report the input value.

pub fn with_initial_text<S: Into<String>>(&mut self, val: S) -> &mut Self

Sets initial text that user can accept or erase.

pub fn default(&mut self, value: T) -> &mut Self

Sets a default.

Out of the box the prompt does not have a default and will continue to display until the user inputs something and hits enter. If a default is set the user can instead accept the default with enter.

pub fn allow_empty(&mut self, val: bool) -> &mut Self

Enables or disables an empty input

By default, if there is no default value set for the input, the user must input a non-empty string.

pub fn show_default(&mut self, val: bool) -> &mut Self

Disables or enables the default value display.

The default behaviour is to append default to the prompt to tell the user what is the default value.

This method does not affect existence of default value, only its display in the prompt!

impl<'a, T> Input<'a, T>

pub fn with_theme(theme: &'a dyn Theme) -> Self

Creates an input prompt with a specific theme.

impl<'a, T> Input<'a, T>where
T: 'a,

pub fn validate_with<V>(&mut self, validator: V) -> &mut Selfwhere
V: Validator<T> + 'a,
V::Err: ToString,

Registers a validator.

Example

let mail: String = Input::new()
    .with_prompt("Enter email")
    .validate_with(|input: &String| -> Result<(), &str> {
        if input.contains('@') {
            Ok(())
        } else {
            Err("This is not a mail address")
        }
    })
    .interact()
    .unwrap();

impl<T> Input<'_, T>where
T: Clone + ToString + FromStr,
<T as FromStr>::Err: Debug + ToString,

pub fn interact_text(&mut self) -> Result<T>

Enables the user to enter a printable ascii sequence and returns the result.

Its difference from interact is that it only allows ascii characters for string, while interact allows virtually any character to be used e.g arrow keys.

The dialog is rendered on stderr.

pub fn interact_text_on(&mut self, term: &Term) -> Result<T>

Like interact_text but allows a specific terminal to be set.

impl<T> Input<'_, T>where
T: Clone + ToString + FromStr,
<T as FromStr>::Err: ToString,

pub fn interact(&mut self) -> Result<T>

Enables user interaction and returns the result.

Allows any characters as input, including e.g arrow keys. Some of the keys might have undesired behavior. For more limited version, see interact_text.

If the user confirms the result is true, false otherwise. The dialog is rendered on stderr.

pub fn interact_on(&mut self, term: &Term) -> Result<T>

Like interact but allows a specific terminal to be set.