Struct dialoguer::MultiSelect

pub struct MultiSelect<'a> { /* private fields */ }

Renders a multi select prompt.

Example usage

use dialoguer::MultiSelect;

let items = vec!["Option 1", "Option 2"];
let chosen : Vec<usize> = MultiSelect::new()
    .items(&items)
    .interact()?;

Implementations

impl<'a> MultiSelect<'a>

pub fn new() -> MultiSelect<'static>

Creates a multi select prompt.

pub fn with_theme(theme: &'a dyn Theme) -> MultiSelect<'a>

Creates a multi select prompt with a specific theme.

pub fn clear(&mut self, val: bool) -> &mut MultiSelect<'a>

Sets the clear behavior of the menu.

The default is to clear the menu.

pub fn defaults(&mut self, val: &[bool]) -> &mut MultiSelect<'a>

Sets a defaults for the menu.

pub fn item<T: ToString>(&mut self, item: T) -> &mut MultiSelect<'a>

Add a single item to the selector.

pub fn item_checked<T: ToString>(
&mut self,
item: T,
checked: bool
) -> &mut MultiSelect<'a>

Add a single item to the selector with a default checked state.

pub fn items<T: ToString>(&mut self, items: &[T]) -> &mut MultiSelect<'a>

Adds multiple items to the selector.

pub fn items_checked<T: ToString>(
&mut self,
items: &[(T, bool)]
) -> &mut MultiSelect<'a>

Adds multiple items to the selector with checked state

pub fn with_prompt<S: Into<String>>(
&mut self,
prompt: S
) -> &mut MultiSelect<'a>

Prefaces the menu with a prompt.

When a prompt is set the system also prints out a confirmation after the selection.

pub fn interact(&self) -> Result<Vec<usize>>

Enables user interaction and returns the result.

The user can select the items with the ‘Space’ bar and on ‘Enter’ the indices of selected items will be returned. The dialog is rendered on stderr. Result contains Vec<index> if user hit ‘Enter’. This unlike interact_opt does not allow to quit with ‘Esc’ or ‘q’.

pub fn interact_opt(&self) -> Result<Option<Vec<usize>>>

Enables user interaction and returns the result.

The user can select the items with the ‘Space’ bar and on ‘Enter’ the indices of selected items will be returned. The dialog is rendered on stderr. Result contains Some(Vec<index>) if user hit ‘Enter’ or None if user cancelled with ‘Esc’ or ‘q’.

pub fn interact_on(&self, term: &Term) -> Result<Vec<usize>>

Like interact but allows a specific terminal to be set.

Examples

 use dialoguer::Select;
 use console::Term;

 fn main() -> std::io::Result<()> {
     let selections = MultiSelect::new()
         .item("Option A")
         .item("Option B")
         .interact_on(&Term::stderr())?;

     println!("User selected options at indices {:?}", selections);

     Ok(())
 }

pub fn interact_on_opt(&self, term: &Term) -> Result<Option<Vec<usize>>>

Like interact_opt but allows a specific terminal to be set.

Examples

use dialoguer::Select;
use console::Term;

fn main() -> std::io::Result<()> {
    let selections = MultiSelect::new()
        .item("Option A")
        .item("Option B")
        .interact_on_opt(&Term::stdout())?;

    match selections {
        Some(positions) => println!("User selected options at indices {:?}", positions),
        None => println!("User exited using Esc or q")
    }

    Ok(())
}