Struct dialoguer::Password

pub struct Password<'a> { /* private fields */ }

Renders a password input prompt.

Example usage

use dialoguer::Password;

let password = Password::new().with_prompt("New Password")
    .with_confirmation("Confirm password", "Passwords mismatching")
    .interact()?;
println!("Length of the password is: {}", password.len());

Implementations

impl Password<'static>

pub fn new() -> Password<'static>

Creates a password input prompt.

impl<'a> Password<'a>

pub fn with_prompt<S: Into<String>>(&mut self, prompt: S) -> &mut Self

Sets the password input prompt.

pub fn report(&mut self, val: bool) -> &mut Self

Indicates whether to report confirmation after interaction.

The default is to report.

pub fn with_confirmation<A, B>( &mut self, prompt: A, mismatch_err: B ) -> &mut Selfwhere A: Into<String>, B: Into<String>,

Enables confirmation prompting.

pub fn allow_empty_password(&mut self, allow_empty_password: bool) -> &mut Self

Allows/Disables empty password.

By default this setting is set to false (i.e. password is not empty).

pub fn validate_with<V>(&mut self, validator: V) -> &mut Selfwhere V: PasswordValidator + 'a, V::Err: ToString,

Registers a validator.

Example

let password: String = Password::new()
    .with_prompt("Enter password")
    .validate_with(|input: &String| -> Result<(), &str> {
        if input.len() > 8 {
            Ok(())
        } else {
            Err("Password must be longer than 8")
        }
    })
    .interact()
    .unwrap();

pub fn interact(&self) -> Result<String>

Enables user interaction and returns the result.

If the user confirms the result is true, false otherwise. The dialog is rendered on stderr.

pub fn interact_on(&self, term: &Term) -> Result<String>

Like interact but allows a specific terminal to be set.

impl<'a> Password<'a>

pub fn with_theme(theme: &'a dyn Theme) -> Self

Creates a password input prompt with a specific theme.