Trait dialoguer::PasswordValidator

pub trait PasswordValidator {
    type Err;

    // Required method
    fn validate(&self, input: &String) -> Result<(), Self::Err>;
}

Trait for password validators.

Required Associated Types

type Err

Required Methods

fn validate(&self, input: &String) -> Result<(), Self::Err>

Invoked with the value to validate.

If this produces Ok(()) then the value is used and parsed, if an error is returned validation fails with that error.