Crate dialoguer

dialoguer is a library for Rust that helps you build useful small interactive user inputs for the command line. It provides utilities to render various simple dialogs like confirmation prompts, text inputs and more.

Best paired with other libraries in the family:

indicatif

console

Crate Contents

Confirmation prompts

Input prompts (regular and password)

Input validation

Selections prompts (single and multi)

Fuzzy select prompt

Other kind of prompts

Editor launching

Crate Features

The following crate features are available:

editor: enables bindings to launch editor to edit strings

fuzzy-select: enables fuzzy select prompt

history: enables input prompts to be able to track history of inputs

password: enables password input prompt

completion: enables ability to implement custom tab-completion for input prompts

By default editor and password are enabled.

Re-exports

pub use console;

Modules

theme

Customizes the rendering of the elements.

Structs

Confirm

Renders a confirm prompt.

Editor

Launches the default editor to edit a string.

Input

Renders an input prompt.

MultiSelect

Renders a multi select prompt.

Password

Renders a password input prompt.

Select

Renders a select prompt.

Sort

Renders a sort prompt.

Traits

Validator

Trait for input validators.