Struct dialoguer::MultiSelect

pub struct MultiSelect<'a> { /* fields omitted */ }

Renders a multi select prompt.

Example usage

use dialoguer::MultiSelect;

let items = vec!["Option 1", "Option 2"];
let chosen : Vec<usize> = MultiSelect::new()
    .items(&items)
    .interact()?;

Implementations

impl<'a> MultiSelect<'a>

pub fn new() -> MultiSelect<'static>

Creates a multi select prompt.

pub fn with_theme(theme: &'a dyn Theme) -> MultiSelect<'a>

Creates a multi select prompt with a specific theme.

pub fn paged(&mut self, val: bool) -> &mut MultiSelect<'a>

Enables or disables paging

pub fn clear(&mut self, val: bool) -> &mut MultiSelect<'a>

Sets the clear behavior of the menu.

The default is to clear the menu.

pub fn defaults(&mut self, val: &[bool]) -> &mut MultiSelect<'a>

Sets a defaults for the menu.

pub fn item<T: ToString>(&mut self, item: T) -> &mut MultiSelect<'a>

Add a single item to the selector.

pub fn item_checked<T: ToString>(
&mut self,
item: T,
checked: bool
) -> &mut MultiSelect<'a>

Add a single item to the selector with a default checked state.

pub fn items<T: ToString>(&mut self, items: &[T]) -> &mut MultiSelect<'a>

Adds multiple items to the selector.

pub fn items_checked<T: ToString>(
&mut self,
items: &[(T, bool)]
) -> &mut MultiSelect<'a>

Adds multiple items to the selector with checked state

pub fn with_prompt<S: Into<String>>(
&mut self,
prompt: S
) -> &mut MultiSelect<'a>

Prefaces the menu with a prompt.

When a prompt is set the system also prints out a confirmation after the selection.

pub fn interact(&self) -> Result<Vec<usize>>

Enables user interaction and returns the result.

The user can select the items with the space bar and on enter the selected items will be returned.

pub fn interact_on(&self, term: &Term) -> Result<Vec<usize>>

Like interact but allows a specific terminal to be set.