Struct dialoguer::MultiSelect

pub struct MultiSelect<'a> { /* private fields */ }

Renders a multi select prompt.

Example

use dialoguer::MultiSelect;

fn main() {
    let items = vec!["foo", "bar", "baz"];

    let selection = MultiSelect::new()
        .with_prompt("What do you choose?")
        .items(&items)
        .interact()
        .unwrap();

    println!("You chose:");

    for i in selection {
        println!("{}", items[i]);
    }
}

Implementations

impl MultiSelect<'static>

pub fn new() -> Self

Creates a multi select prompt with default theme.

impl MultiSelect<'_>

pub fn clear(self, val: bool) -> Self

Sets the clear behavior of the menu.

The default is to clear the menu.

pub fn defaults(self, val: &[bool]) -> Self

Sets a defaults for the menu.

pub fn max_length(self, val: usize) -> Self

Sets an optional max length for a page

Max length is disabled by None

pub fn item<T: ToString>(self, item: T) -> Self

Add a single item to the selector.

pub fn item_checked<T: ToString>(self, item: T, checked: bool) -> Self

Add a single item to the selector with a default checked state.

pub fn items<T: ToString>(self, items: &[T]) -> Self

Adds multiple items to the selector.

pub fn items_checked<T: ToString>(self, items: &[(T, bool)]) -> Self

Adds multiple items to the selector with checked state

pub fn with_prompt<S: Into<String>>(self, prompt: S) -> Self

Prefaces the menu with a prompt.

By default, when a prompt is set the system also prints out a confirmation after the selection. You can opt-out of this with report.

pub fn report(self, val: bool) -> Self

Indicates whether to report the selected values after interaction.

The default is to report the selections.

pub fn interact(self) -> Result<Vec<usize>>

Enables user interaction and returns the result.

The user can select the items with the ‘Space’ bar and on ‘Enter’ the indices of selected items will be returned. The dialog is rendered on stderr. Result contains Vec<index> if user hit ‘Enter’. This unlike interact_opt does not allow to quit with ‘Esc’ or ‘q’.

pub fn interact_opt(self) -> Result<Option<Vec<usize>>>

Enables user interaction and returns the result.

The user can select the items with the ‘Space’ bar and on ‘Enter’ the indices of selected items will be returned. The dialog is rendered on stderr. Result contains Some(Vec<index>) if user hit ‘Enter’ or None if user cancelled with ‘Esc’ or ‘q’.

Example

use dialoguer::MultiSelect;

fn main() {
    let items = vec!["foo", "bar", "baz"];

    let ordered = MultiSelect::new()
        .items(&items)
        .interact_opt()
        .unwrap();

    match ordered {
        Some(positions) => {
            println!("You chose:");

            for i in positions {
                println!("{}", items[i]);
            }
        },
        None => println!("You did not choose anything.")
    }
}

pub fn interact_on(self, term: &Term) -> Result<Vec<usize>>

Like interact but allows a specific terminal to be set.

pub fn interact_on_opt(self, term: &Term) -> Result<Option<Vec<usize>>>

Like interact_opt but allows a specific terminal to be set.

impl<'a> MultiSelect<'a>

pub fn with_theme(theme: &'a dyn Theme) -> Self

Creates a multi select prompt with a specific theme.

Example

use dialoguer::{theme::ColorfulTheme, MultiSelect};

fn main() {
    let selection = MultiSelect::with_theme(&ColorfulTheme::default())
        .items(&["foo", "bar", "baz"])
        .interact()
        .unwrap();
}