Struct dialoguer::Sort

pub struct Sort<'a> { /* private fields */ }

Renders a sort prompt.

Returns list of indices in original items list sorted according to user input.

Example usage

use dialoguer::Sort;

let items_to_order = vec!["Item 1", "Item 2", "Item 3"];
let ordered = Sort::new()
    .with_prompt("Order the items")
    .items(&items_to_order)
    .interact()?;

Implementations

impl Sort<'static>

pub fn new() -> Self

Creates a sort prompt.

impl Sort<'_>

pub fn clear(&mut self, val: bool) -> &mut Self

Sets the clear behavior of the menu.

The default is to clear the menu after user interaction.

pub fn max_length(&mut self, val: usize) -> &mut Self

Sets an optional max length for a page

Max length is disabled by None

pub fn item<T: ToString>(&mut self, item: T) -> &mut Self

Add a single item to the selector.

pub fn items<T: ToString>(&mut self, items: &[T]) -> &mut Self

Adds multiple items to the selector.

pub fn with_prompt<S: Into<String>>(&mut self, prompt: S) -> &mut Self

Prefaces the menu with a prompt.

By default, when a prompt is set the system also prints out a confirmation after the selection. You can opt-out of this with report.

pub fn report(&mut self, val: bool) -> &mut Self

Indicates whether to report the selected order after interaction.

The default is to report the selected order.

pub fn interact(&self) -> Result<Vec<usize>>

Enables user interaction and returns the result.

The user can order the items with the ‘Space’ bar and the arrows. On ‘Enter’ ordered list of the incides of items will be returned. The dialog is rendered on stderr. Result contains Vec<index> if user hit ‘Enter’. This unlike interact_opt does not allow to quit with ‘Esc’ or ‘q’.

pub fn interact_opt(&self) -> Result<Option<Vec<usize>>>

Enables user interaction and returns the result.

The user can order the items with the ‘Space’ bar and the arrows. On ‘Enter’ ordered list of the incides of items will be returned. The dialog is rendered on stderr. Result contains Some(Vec<index>) if user hit ‘Enter’ or None if user cancelled with ‘Esc’ or ‘q’.

pub fn interact_on(&self, term: &Term) -> Result<Vec<usize>>

Like interact but allows a specific terminal to be set.

Examples

 use dialoguer::Sort;
 use console::Term;

 fn main() -> std::io::Result<()> {
     let selections = Sort::new()
         .item("Option A")
         .item("Option B")
         .interact_on(&Term::stderr())?;

     println!("User sorted options as indices {:?}", selections);

     Ok(())
 }

pub fn interact_on_opt(&self, term: &Term) -> Result<Option<Vec<usize>>>

Like interact_opt but allows a specific terminal to be set.

Examples

use dialoguer::Sort;
use console::Term;

fn main() -> std::io::Result<()> {
    let selections = Sort::new()
        .item("Option A")
        .item("Option B")
        .interact_on_opt(&Term::stdout())?;

    match selections {
        Some(positions) => println!("User sorted options as indices {:?}", positions),
        None => println!("User exited using Esc or q")
    }

    Ok(())
}

impl<'a> Sort<'a>

pub fn with_theme(theme: &'a dyn Theme) -> Self

Creates a sort prompt with a specific theme.