Struct dialoguer::theme::ColorfulTheme

pub struct ColorfulTheme {

    pub defaults_style: Style,
    pub prompt_style: Style,
    pub prompt_prefix: StyledObject<String>,
    pub prompt_suffix: StyledObject<String>,
    pub success_prefix: StyledObject<String>,
    pub success_suffix: StyledObject<String>,
    pub error_prefix: StyledObject<String>,
    pub error_style: Style,
    pub hint_style: Style,
    pub values_style: Style,
    pub active_item_style: Style,
    pub inactive_item_style: Style,
    pub active_item_prefix: StyledObject<String>,
    pub inactive_item_prefix: StyledObject<String>,
    pub checked_item_prefix: StyledObject<String>,
    pub unchecked_item_prefix: StyledObject<String>,
    pub picked_item_prefix: StyledObject<String>,
    pub unpicked_item_prefix: StyledObject<String>,

}

A colorful theme

Fields
defaults_style: Style

The style for default values

prompt_style: Style

The style for prompt

prompt_prefix: StyledObject<String>

Prompt prefix value and style

prompt_suffix: StyledObject<String>

Prompt suffix value and style

success_prefix: StyledObject<String>

Prompt on success prefix value and style

success_suffix: StyledObject<String>

Prompt on success suffix value and style

error_prefix: StyledObject<String>

Error prefix value and style

error_style: Style

The style for error message

hint_style: Style

The style for hints

values_style: Style

The style for values on prompt success

active_item_style: Style

The style for active items

inactive_item_style: Style

The style for inactive items

active_item_prefix: StyledObject<String>

Active item in select prefix value and style

inactive_item_prefix: StyledObject<String>

Inctive item in select prefix value and style

checked_item_prefix: StyledObject<String>

Checked item in multi select prefix value and style

unchecked_item_prefix: StyledObject<String>

Unchecked item in multi select prefix value and style

picked_item_prefix: StyledObject<String>

Picked item in sort prefix value and style

unpicked_item_prefix: StyledObject<String>

Unpicked item in sort prefix value and style