Trait dialoguer::InputValidator

pub trait InputValidator<T> {
    type Err;

    // Required method
    fn validate(&mut self, input: &T) -> Result<(), Self::Err>;
}

Trait for input validators.

A generic implementation for Fn(&str) -> Result<(), E> is provided to facilitate development.

Required Associated Types

type Err

Required Methods

fn validate(&mut self, input: &T) -> Result<(), Self::Err>

Invoked with the value to validate.

If this produces Ok(()) then the value is used and parsed, if an error is returned validation fails with that error.